+++ kiteffmpeg/src/jvmAndAndroidMain/kotlin/io/github/yuroyami/kiteffmpeg/Transcoder.jvm.kt
@@ -2,8 +2,10 @@ package io.github.yuroyami.kiteffmpeg
 
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
+import kotlin.coroutines.cancellation.CancellationException
 
 public actual object Transcoder {
+    @Throws(FFmpegException::class, CancellationException::class)
     public actual suspend fun transcode(
         input: String,
         output: String,